fix: remove isaac from binary header
We will not be keeping encrypted packets in the blob as there is no
actual benefit to doing so. Every packet must be correctly spliced
by the proxy anyway, there's no argument for something being corrupt.
If it was corrupt, it would be unusable regardless. As such, in order
to simplify usage all around, decrypt it before passing it into
any decoders.