feat: make use of cache for npc info decoding
This fixes a potential decoder roadblock where head or body recol/retex
could be used for npcs, which requires knowing the length of an array
in npc type within the cache to continue. Without this information,
npc info cannot continue processing - not this, nor the future packets.
This is because npc info is stateful and lack of state can result in
future packets failing to decode too.