chore: make 'proxy' builds rebuild plugins/transcribers
This will result in any decoder and transcriber plugins being copied
over so that they can be used during various tasks.
Multiple people were hitting the roadblock of plugins not being found
and not knowing they just needed to build the entire repository.
The downside to this approach is that every execution will now result
in the compilation and copying of all the plugins that exist, but for
now, this shouldn't be a problem as there aren't many of them and it
takes about a second to do.