feat: js5 group downloader
Adds a simple single-threaded blocking JS5 group downloader.
This downloader will request one group at a time, waiting for the
response before continuing with the next. As such, it is not well suited
for downloading large amounts of files. It is more than enough for our
needs here, as we will only request a couple groups anyway.