feat: add support for exporting binary credentials
These credentials are used to uniquely identify binaries created
by a specific user. These cannot be used to gain access to peoples'
accounts. They are strictly for the public archive in the future!

By using SHA-256 to hash the user id and user hash values, we can
come to the same 'accountHash' value that is stored in the binary
headers. It is not possible to get the two components that make up
a binary out of the hash, it's a one-way conversion. The public
archive will require inputting the two components, after which the
archive will perform the SHA-256 hashing on it, which will result
in all the binaries associated with that hash being listed.