fix: check for necessary vcomp140 library when launching native client
As someone that just reinstalled windows and didn't launch the native
client via the Jagex Launcher, I lacked the necessary Visual C++
Redistributable and its libraries, and thus was unable to boot the
native client. This commit gives a clean error indicating what went
wrong if that were to happen.