fix: make .txt transcripts of bin differ in last-modified timestamp
This ensures that tools such as notepad will pick up on changes to the
file, as they rely on the last-modified date changing.
The new implementation will begin the timestamp at .bin +1ms,
incrementing by +1ms every time the tool is re-ran, if the .txt file
already exists. Visually the timestamps will remain the same, but
for sorting purposes, and file-change-listeners, it'll behave as
one would want.