fix: log start/end fill if time is logged on headbars
There are some unusual usages in Old School, where startFill and
endFill are the same value, but timings are used. The timings have
no effect in such cases, so it creates a confusing log where you only
end up with "fill", "starttime" and "endtime" properties, which
doesn't make any sense. It'll now log the start and end fill to
get the point across a little better, even if the values equal.