feat: add support for preferring official dictionary over custom
Allows for us to use Jagex dictionaries (gamevals) wherever they can
be used, but fall back to our custom dictionaries any other time.

The benefit of this is that if a server has fallen behind and an
id has shifted, it won't suddenly be reporting incorrectly - we'd
still receive the correct name, just with a different id, which means
that if we update our server that makes use of gamevals, we can
automatically sort out the ids and keep the content working as intended.