refactor: move certain npc info settings to filters
This change removes the "Npc Removal" and "Hide Inactive NPCs"
settings, and introduces them as as individual filters:
Note: These filters are checked as "are any of them enabled".
An entry is logged if at least one is enabled, even if some part of
it would otherwise be disabled.

1. Npc Add - Always logs NPCs which were just added to high resolution,
regardless of their extended info and movement status.
2. Npc Delete - Always logs NPCs which were just deleted from high
resolution, regardless of the other status.
3. Npc Movement - Always logs an NPC that crawled, walked, ran or
teleported.
4. Npc Ext. Info - A group-level toggle for extended info. This one is
a little different from the rest. If enabled, only the individual ext
info which the user has selected will end up logged. This will forcibly
log NPCs even if they were added and the npc-add filter is turned off,
similarly it'll forcibly log any NPC which moved with movement off.